# docs/swagger.yaml
basePath: /
definitions:
...
  models.Branch:
    properties:
      aashram_area:
        minimum: 0
        type: number
      address:
        maxLength: 500
        type: string
      branch_members:
        items:
          $ref: '#/definitions/models.BranchMember'
        type: array
      children:
        items:
          $ref: '#/definitions/models.Branch'
        type: array
      city:
        $ref: '#/definitions/models.City'
      city_id:
        minimum: 1
        type: integer
      contact_number:
        maxLength: 20
        type: string
      coordinator_name:
        maxLength: 255
        minLength: 2
        type: string
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        minimum: 1
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      daily_end_time:
        type: string
      daily_start_time:
        type: string
      district:
        $ref: '#/definitions/models.District'
      district_id:
        minimum: 1
        type: integer
      email:
        maxLength: 255
        type: string
      established_on:
        type: string
      id:
        type: integer
      infrastructure:
        items:
          $ref: '#/definitions/models.BranchInfrastructure'
        type: array
      name:
        maxLength: 255
        minLength: 2
        type: string
      open_days:
        maxLength: 100
        type: string
      parent:
        $ref: '#/definitions/models.Branch'
      parent_branch_id:
        type: integer
      pincode:
        type: string
      police_station:
        maxLength: 100
        type: string
      post_office:
        maxLength: 100
        type: string
      state:
        $ref: '#/definitions/models.State'
      state_id:
        minimum: 1
        type: integer
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - contact_number
    - name
    type: object
  models.BranchInfrastructure:
    properties:
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        minimum: 1
        type: integer
      count:
        minimum: 0
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      id:
        type: integer
      type:
        maxLength: 100
        minLength: 2
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - branch_id
    - count
    - type
    type: object
  models.BranchMember:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        minimum: 1
        type: integer
      branch_role:
        maxLength: 100
        type: string
      created_by:
        type: string
      created_on:
        type: string
      date_of_birth:
        type: string
      date_of_samarpan:
        type: string
      id:
        type: integer
      member_type:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      qualification:
        maxLength: 255
        type: string
      responsibility:
        maxLength: 500
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - branch_id
    - member_type
    - name
    type: object
  models.City:
    properties:
      id:
        type: integer
      name:
        type: string
      state_id:
        type: integer
    type: object
  models.Country:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.District:
    properties:
      country_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      state_id:
        type: integer
    type: object
...
  models.Event:
    properties:
      event_name:
        type: string
      id:
        type: integer
    type: object
...
  models.State:
    properties:
      country_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Volunteer:
    properties:
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        minimum: 1
        type: integer
      contact:
        maxLength: 20
        type: string
      created_by:
        type: string
      created_on:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        minimum: 1
        type: integer
      id:
        type: integer
      mention_seva:
        maxLength: 500
        minLength: 2
        type: string
      number_of_days:
        maximum: 365
        minimum: 0
        type: integer
      seva_involved:
        maxLength: 500
        minLength: 2
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
      volunteer_name:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - branch_id
    - event_id
    - volunteer_name
    type: object
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
...
  /api/events/{event_id}/volunteers:
    get:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Volunteer'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get volunteers by event ID
      tags:
      - Volunteers
...
  /api/volunteers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Volunteer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all volunteers
      tags:
      - Volunteers
    post:
      consumes:
      - application/json
      description: Store volunteer details
      parameters:
      - description: Volunteer payload
        in: body
        name: volunteer
        required: true
        schema:
          $ref: '#/definitions/models.Volunteer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a volunteer
      tags:
      - Volunteers
  /api/volunteers/{id}:
    delete:
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a volunteer
      tags:
      - Volunteers
    put:
      consumes:
      - application/json
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a volunteer
      tags:
      - Volunteers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"